+++ pkgs/swift2objc/lib/src/transformer/_core/utils.dart
@@ -3,9 +3,12 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import '../../ast/_core/interfaces/declaration.dart';
+import '../../ast/_core/shared/parameter.dart';
 import '../../ast/_core/shared/referred_type.dart';
-import '../../ast/declarations/built_in/built_in_declaration.dart';
 import '../../ast/declarations/compounds/class_declaration.dart';
+import '../../ast/declarations/compounds/members/initializer_declaration.dart';
+import '../../ast/declarations/compounds/members/property_declaration.dart';
+import '../../transformer/_core/primitive_wrappers.dart';
 import '../transform.dart';
 import 'unique_namer.dart';
 
@@ -78,4 +81,25 @@ import 'unique_namer.dart';
   } else {
     throw UnimplementedError('Unknown type: $type');
   }
+}
+
+InitializerDeclaration buildWrapperInitializer(
+  PropertyDeclaration wrappedClassInstance,
+) {
+  return InitializerDeclaration(
+    id: '',
+    params: [
+      Parameter(
+        name: '_',
+        internalName: 'wrappedInstance',
+        type: wrappedClassInstance.type,
+      )
+    ],
+    isOverriding: false,
+    isFailable: false,
+    throws: false,
+    async: false,
+    statements: ['self.${wrappedClassInstance.name} = wrappedInstance'],
+    hasObjCAnnotation: wrappedClassInstance.hasObjCAnnotation,
+  );
 }
