+++ pkgs/swift2objc/lib/src/transformer/transformers/transform_compound.dart
@@ -5,7 +5,6 @@
 import '../../ast/_core/interfaces/compound_declaration.dart';
 import '../../ast/_core/interfaces/declaration.dart';
 import '../../ast/_core/interfaces/nestable_declaration.dart';
-import '../../ast/_core/shared/parameter.dart';
 import '../../ast/declarations/built_in/built_in_declaration.dart';
 import '../../ast/declarations/compounds/class_declaration.dart';
 import '../../ast/declarations/compounds/members/initializer_declaration.dart';
@@ -13,6 +12,7 @@ import '../../ast/declarations/compounds/members/method_declaration.dart';
 import '../../ast/declarations/compounds/members/property_declaration.dart';
 import '../../parser/_core/utils.dart';
 import '../_core/unique_namer.dart';
+import '../_core/utils.dart';
 import '../transform.dart';
 import 'transform_function.dart';
 import 'transform_initializer.dart';
@@ -97,25 +97,4 @@ ClassDeclaration transformCompound(
     ..sort((Declaration a, Declaration b) => a.id.compareTo(b.id));
 
   return transformedCompound;
-}
-
-InitializerDeclaration buildWrapperInitializer(
-  PropertyDeclaration wrappedClassInstance,
-) {
-  return InitializerDeclaration(
-    id: '',
-    params: [
-      Parameter(
-        name: '_',
-        internalName: 'wrappedInstance',
-        type: wrappedClassInstance.type,
-      )
-    ],
-    isOverriding: false,
-    isFailable: false,
-    throws: false,
-    async: false,
-    statements: ['self.${wrappedClassInstance.name} = wrappedInstance'],
-    hasObjCAnnotation: wrappedClassInstance.hasObjCAnnotation,
-  );
 }
